fix(session-ingest): add fetch() to SessionIngestRPC for service binding HTTP calls
The service binding from cloud-agent-next to session-ingest uses
entrypoint: SessionIngestRPC. RPC calls work, but .fetch() failed with
'Handler does not export a fetch() function' because the entrypoint had
no fetch method — the Hono app lived on the default export.

- Extract Hono app to app.ts to avoid circular imports
- Add fetch() to SessionIngestRPC that delegates to the Hono app
- Revert cloud-agent-next URL back to service binding (https://session-ingest/...)
